Extract `selectTemplateIdx` to own file
